requirements.txt
streamlit
pandas
requests
reportlab
google-genai
